Switched to ELF for .lac files

+++ lavm/src/system_calls/entry_point.h
@@ -1,3 +1,3 @@
 #include "cpu.h"
 
-uint32_t la_system_call(struct cpu * cpu, uint8_t syscall_number, uint32_t arg);
+uint64_t la_system_call(struct cpu * cpu, uint8_t syscall_number, uint64_t arg);
